requirements.txt
yfinance
pandas
feedparser
flask
flask-cors
python-dotenv
requests
